web/Web/requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi


# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms


# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
#wheel==0.29.0


# Forms
django-braces==1.14.0

# Models
django-model-utils==4.0.0


# Python-PostgreSQL Database Adapter
psycopg2==2.8.5 --no-binary psycopg2

# Unicode slugification
#awesome-slugify==1.6.5

# Time zones support
#pytz==2017.2

# Redis support
django-redis==4.11.0  # https://github.com/jazzband/django-redis
redis==3.5.0  # https://github.com/andymccurdy/redis-py


# Your custom requirements go here
httplib2
xmltodict
lxml
raven
uwsgi
beautifulsoup4==4.4.0
django-autocomplete-light==3.5.1
django-notifications-hq==1.6.0
